Refactor extra code
