+++ crates/module-system/hyperlane/tests/integration/with_agent/helpers/evm.rs
@@ -9,7 +9,7 @@ use sha3::{Digest, Keccak256};
 use sov_hyperlane_integration::{EthAddress, Message};
 use sov_modules_api::macros::config_value;
 use sov_modules_api::{Amount, HexHash, HexString};
-use sov_test_utils::docker::{print_logs_from_container, pull_image_with_retries};
+use sov_test_utils::docker::{pre_pull_image_with_auth_fallback, print_logs_from_container};
 use std::path::PathBuf;
 use std::sync::atomic::{AtomicU64, Ordering};
 use std::time::{Duration, Instant};
@@ -32,26 +32,7 @@ pub struct AnvilRunner {
 impl AnvilRunner {
     pub async fn new() -> Self {
         tracing::info!("Starting anvil container...");
-        if let Err(err) = pull_image_with_retries(GenericImage::new(ANVIL_IMAGE, TAG)).await {
-            let err_text = err.to_string();
-            let auth_failure = err_text.contains("status code 401")
-                || err_text.contains("status code 403")
-                || err_text.contains("denied");
-
-            if auth_failure {
-                tracing::warn!(
-                    %err,
-                    image = ANVIL_IMAGE,
-                    tag = TAG,
-                    "Pre-pull failed with registry authorization response; deferring image resolution to Docker on container start"
-                );
-            } else {
-                panic!(
-                    "failed to pull anvil image {ANVIL_IMAGE}:{TAG}: {err}. \
-                     Hint: verify GHCR connectivity or pre-pull the image before tests."
-                );
-            }
-        }
+        pre_pull_image_with_auth_fallback(GenericImage::new(ANVIL_IMAGE, TAG)).await;
 
         // Hard code tag, so we don't accidental breakages
         let (container, loaded_state) = {
