+++ crates/module-system/hyperlane/tests/integration/with_agent/helpers/hyperlane_cli.rs
@@ -5,7 +5,7 @@ use crate::with_agent::configs::{
 use crate::with_agent::helpers::{parse_eth_addr, DEPLOYER_ACCOUNT, EVM_MAILBOX, RELAYER_ACCOUNT};
 use sov_hyperlane_integration::EthAddress;
 use sov_modules_api::HexHash;
-use sov_test_utils::docker::pull_image_with_retries;
+use sov_test_utils::docker::pre_pull_image_with_auth_fallback;
 use std::collections::HashMap;
 use std::path::Path;
 use std::str::FromStr;
@@ -213,23 +213,7 @@ fn prepare_core_deploy_data(data_path: &Path, anvil_port: u16, host_address: &st
 
 // Waits for some time while hyperlane-cli exit with status code 0
 async fn wait_till_container_exit(hyperlane_cli_image: ContainerRequest<GenericImage>) -> String {
-    if let Err(err) = pull_image_with_retries(GenericImage::new(IMAGE, TAG)).await {
-        let err_text = err.to_string();
-        let auth_failure = err_text.contains("status code 401")
-            || err_text.contains("status code 403")
-            || err_text.contains("denied");
-
-        if auth_failure {
-            tracing::warn!(
-                %err,
-                image = IMAGE,
-                tag = TAG,
-                "Pre-pull failed with registry authorization response; deferring image resolution to Docker on container start"
-            );
-        } else {
-            panic!("Failed to pull hyperlane-cli image: {err}");
-        }
-    }
+    pre_pull_image_with_auth_fallback(GenericImage::new(IMAGE, TAG)).await;
 
     let container: testcontainers::ContainerAsync<GenericImage> = hyperlane_cli_image
         .start()
